refactor: architecture, types, imports, styles
