+++ src/components/resultField/resultField.tsx
@@ -1,6 +1,6 @@
-import {Card}from '../card/card';
+import {Card}from '@components/card/card';
 import {usePagination} from '@utils/hooks/usePagination';
-import './resultField.css';
+import resultFieldStyles from "./resultField.module.css";
 
 type ResultFieldProps = {
   movies: object[];
@@ -9,12 +9,12 @@ type ResultFieldProps = {
 export const ResultField: React.FC<ResultFieldProps> = ({ movies }) => {
   const { currentMovies, currentPage, nextPage, prevPage } = usePagination(movies);
   return (
-    <div className="resultContainer">
-      <div className="resultField">
+    <div className={resultFieldStyles.resultContainer}>
+      <div className={resultFieldStyles.resultField}>
         <Card currentMovies={currentMovies} />
       </div>
       {movies.length > 6 && (
-        <div className="pagination">
+        <div className={resultFieldStyles.pagination}>
           <button onClick={prevPage} disabled={currentPage === 1}>
             &#60;
           </button>
