+++ src/pages/movieFinder/movieFInder.module.css
@@ -1,44 +1,10 @@
-.movieFinder {
-    min-width: 827px;
-}
-
-.title {
-    font-size: 80px;
-    line-height: 120px;
-    color: #ADFF2F;
-}
-
-.information {
-    font-size: 36px;
-    line-height: 54px;
-    color: #FFFFFF;
-}
-
-.signature {
-    text-align: left;
-    font-size: 18px;
-    line-height: 27px;
-    color: #BDBDBD;
-}
-
-.form {
-    display: flex;
-    flex-direction: row;
-}
-
-.searchField {
-    flex-grow: 1;
-    border-radius: 10px 0 0 10px;
-    font-size: 24px;
-    line-height: 36px;
-    color: #000000;
-    padding-left: 33px;
-}
-
-.submit {
-    border-radius: 0 10px 10px 0;
-    font-size: 24px;
-    line-height: 36px;
-    background-color: var(--button-bg-color);
-    color: var(--button-color);
+.container {
+  --background-color: white;
+	--display: flex;
+	--justify-content: center;
+	--justify-content: center;
+	--font-family:'Poppins', sans-serif;
+	--color: rgba(0, 0, 0, 1);
+	--font-weight: 400;
+	--letter-spacing: 0%;
 }
