--- src/pages/movieFinder/movieFinder.css
@@ -1,10 +1,0 @@
-:root {
-  --background-color: white;
-	--display: flex;
-	--justify-content: center;
-	--justify-content: center;
-	--font-family:'Poppins', sans-serif;
-	--color: rgba(0, 0, 0, 1);
-	--font-weight: 400;
-	--letter-spacing: 0%;
-}
