+++ src/pages/movieFinder/movieFinder.tsx
@@ -1,13 +1,12 @@
 import React, { useState } from 'react';
-import {InputField }from './components/inputField/inputField';
-import {ResultField} from './components/resultField/resultField';
-import './components/inputField/inputFiels.css';
-import './components/resultField/resultField.css';
-import './movieFinder.css'
+import {InputField }from '@components/inputField/inputField';
+import {ResultField} from '@components/resultField/resultField';
+import movieFinderStyles from "./movieFInder.module.css";
+import { Movie } from '@utils/types/types';
 
 export const MovieFinder: React.FC = () => {
   const [searchTerm, setSearchTerm] = useState<string>('');
-  const [movies, setMovies] = useState<any[]>([]);
+  const [movies, setMovies] = useState<Movie[]>([]);
   const fetchMovies = async (query: string) => {
     try {
       const response = await fetch(`https://www.omdbapi.com/?apikey=3e3e7f8f&s=${query}`);
@@ -31,7 +30,7 @@ export const MovieFinder: React.FC = () => {
   };
 
   return (
-    <div>
+    <div className={movieFinderStyles.container}>
       <InputField
 				searchTerm={searchTerm}
         handleInputChange={handleInputChange}
